+++ generators/app/tasks/appSetup/coreFiles/packageJsonScripts.js
@@ -21,9 +21,8 @@ module.exports = function packgeJsonScripts() {
   packageJson.scripts.lint = 'eslint src --ext .js,.ts,.jsx,.tsx';
   packageJson.scripts['lint-fix'] = 'eslint src --ext .js,.ts,.jsx,.tsx --fix';
   packageJson.scripts['lint-diff'] =
-    'git diff --name-only --cached --no-error-on-unmatched-pattern --relative | grep -E "\\.(ts|tsx|js|jsx)$" | xargs eslint';
+    'git diff --name-only --relative --diff-filter=ACM | grep -E "\\.(ts|tsx|js|jsx)$" | xargs eslint';
   packageJson.scripts.tsc = 'tsc';
-
   packageJson.husky = packageJson.husky || {};
   packageJson.husky.hooks = packageJson.husky.hooks || {};
   packageJson.husky.hooks['pre-commit'] = 'yarn run lint-diff && yarn run tsc';