+++ generators/app/tasks/appSetup/coreFiles/androidProjectSetup.js
@@ -30,7 +30,7 @@ function updateAppBuildGradle() {
   );
   updatedBuildGradleContent = updatedBuildGradleContent.replace(
     'if (enableHermes) {',
-    "implementation 'androidx.appcompat:appcompat:1.2.0-alpha03'\n\timplementation 'androidx.multidex:multidex:2.0.1'\n\tif (enableHermes) {"
+    "implementation 'androidx.appcompat:appcompat:1.3.0-alpha02'\n\timplementation 'androidx.multidex:multidex:2.0.1'\n\tif (enableHermes) {"
   );
   this.fs.write(`${this.projectName}/android/app/build.gradle`, updatedBuildGradleContent);
 }