+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Replaced `react-native` command with `npx react-native ...` (RN community indication)
 - Updated again some Firebase configuration (App, Crashlytics and Performance)
 - Fixed `lint-diff` script in Generator and Bootstraped Project
+- Updated `androidx.appcompat` dependency in Android project configuration
 
 ## [1.3.0] - 2020-09-02